Vectorization cleanup (#393)
Currently, vectorization analysis is done by first calling
`getInnerDimVectorizableWidth` to compute teh vectorization factor of
the innermost dimension, then expand that to contiguous merge of
`getMaxVectorizableWidth`. However, these two functions are very
similar, almost a copy-paste of each other. Therefore, it makes no sense
to use `getInnerDimVectorizableWidth` then `getMaxVectorizableWidth`, we
should remove `getInnerDimVectorizableWidth` and go directly to
`getMaxVectorizableWidth`.

This PR removes `getInnerDimVectorizableWidth`, anded a boolean option
`contig_merge` to `getMaxVectorizableWidth` to reproduce the behavior of
`getInnerDimVectorizableWidth`. The `contig_merge=false` is only used in
the transpose scheduler, where the which dimensions are contiguously
merge to which ID is more complicated and is not considered in
heuristics.

Unfortunately, because the variable name change, I can not diff the
kernel to check if they match exactly or not, but I do briefly skimmed
through the diff by eye, and I don't find any change in the
vectorization factor.

The V100 failure is OOM.