Add vector of scalars abstraction to Python Frontend structures (#199)
This PR is mostly dedicated to adding the new `StateType` of `Vector` to
the `State` object found in the `recording_state_` of the
`FusionDefinition`. I moved the `recording_state_` to the `FusionState`
because it was meant to be stored in tandem with the Fusion Record
`recording_`.

So this PR does a few things:

1. Adds `Vector` to `StateType` in the `serde` `fbs` file. That was
regenerated to modify the this file
[csrc/serde/python_fusion_cache_generated.cpp](https://github.com/NVIDIA/Fuser/compare/main...add_vector_of_scalars#diff-83b37fb90b52ea1008960f5854975869584e4fe345bb8ae204a0348c178d4192)
2. Adds a `Vector` object like `Tensor` and `Scalar` in
`fusion_definition.h`.
3. The object `recording_state_` that stores `State` objects discovered
while creating records was moved from the `FusionDefinition` to the
`FusionState`. This change wasn't needed but I did it because I was
considering adding asserts based on the `StateType` but I ended up not
doing that.
4. Added new methods to the `FusionState` to handle a
`std::vector<Val*>` which I refer to as Vector of Scalars.

1/3 Stacked PRs.

TODO:

- [x]  Fix up printing of Vector of State

```
index dc168c4..ba53b59 100644
--- a/csrc/python_frontend/fusion_state.cpp
+++ b/csrc/python_frontend/fusion_state.cpp
@@ -29,6 +29,8 @@ std::ostream& operator<<(std::ostream& os, const State& state) {
     os << "S";
   } else if (state.stype == serde::StateType_Tensor) {
     os << "T";
+  } else if (state.stype == serde::StateType_Vector) {
+    os << "V";
   } else if (state.stype == serde::StateType_None) {
     os << "None";
   } else {
```