Fix collisions in cache system when device != 0 and no tensor inputs (#404)
Last night's CI failed with a case where a Fusion was not compiled in
the Python test `test_selected_device`:

https://gitlab-master.nvidia.com/dl/pytorch/fuser-gh-mirror/-/jobs/61110157

This happened because the fusion has no tensor inputs but was asked to
run on a separate device from where it was first compiled, and did not
recognize that it needed to be recompiled.

In a similar case in testing integration, I've seen the wrong kernel
sometimes returned by `getKernelRuntimeFor`, resulting in outputs that
lie on the wrong device.

The reason this happens is that until now, we printed tensor input
devices when generating a `KernelArgumentsHolder` cache ID, and relied
on that to determine the device we're running on. As of 6915258 we are
able to instruct NVFuser to run these on separate devices, but the cache
ID was not properly updated to reflect this. This PR corrects that
oversight.

Additionally, when checking whether we need to compile a
`FusionKernelRuntime` we previously made a separate call to
`getKernelRuntimeFor` and checked if it was compiled. Instead, this PR
changes it to just directly check whether the previously-retrieved
runtime is compiled. It also adds a `device` arg to
`FusionKernelRuntime::isCompiled` so that these can be checked without
directly retrieving them (although I'm not sure when that would be
needed).