Add `define_vector` function to `FusionDefinition` in Python API (#219)
Adding the `fd.define_vector()` function to the `FusionDefinition`. 

This includes the underlying `VectorRecord` but does not include the
slightly different record for consuming `Scalar`s and producing a
`Vector`. This will occur in a future PR.
1. `fd.define_vector(size)` -- allows you to define an input vector
which is really just a bunch of `Scalar`s.
2. `fd.define_vector([2, 2, 3])` -- allows you to define a constant
`Vector` of integers.

I did not bump this `version` because this PR is not meant to be used,
by itself, but does not hurt any other functionality.