Add all symbolic tensor extents to leaf_dynamic_vals_
This might make the lookups there a bit slower, as previously these
would be very small sets. However, this is necessary in order to find
differences in empty-tensor concretization. Note that the actual keys
won't commonly change but this will find input vals that affect output
sizes, such as in FusionStandaloneIota_CUDA, which was failing before
this fix.