Cleanups to FusionExecutorCache and DynamicFusionConcretizationInfo (#466)
This PR makes a few changes to the dynamic Fusion machinery. Mainly,
this is to make ownership of `DynamicTransformInitialInfo`,
`DynamicTransformConcretizationInfo`, and `ExpressionEvaluator` more
clear. It avoids cloning (and using fusion-managed data for)
`DynamicTransformConcretizationInfo` and does not store `Statement`
pointers in that object. Instead we refer to lists of dynamic ops that
are held in `initial_info`, which _can_ be cloned and is managed by the
`Fusion`. This makes it easier to compare concretization infos that
apply to different `Fusion`s, and to reuse it for the actual
concretization without needing to clone again.

Minor changes:
- Move iteration through root dynamic vals to `finalizeDynamicVals()`.
- Use `executor_utils::bindInputs()` instead of manually binding inputs
as in `getConcretizationInfo()`.
- Expose `FusionExecutorCache::countConcretizations()` and
`FusionExecutorCache::countRuntimes()`