Add tensor factory `torch.tensor` (#468)
[`torch.tensor`](https://pytorch.org/docs/stable/generated/torch.tensor.html?highlight=tensor#torch.tensor)
is a tensor factory that constructs a tensor from given scalar values.
This PR adds support for this tensor factory to our C++ API. Python
frontend is not supported yet, because of the following reason:

This tensor factory is not something we prioritize to have. However,
building this out requires improving our array support in codegen. The
main purpose of this PR is to build out array infrastructure and a great
portion of this work will be reused for TMA support (which takes an
argument of N dimensional coordinates as index). The prototype for TMA
is at https://github.com/NVIDIA/Fuser/pull/455

In the future, if this tensor factory turns out to be useful, we can add
python frontend at that time. (It will NOT be easy, because python's
`[[1, 2], [2, 4]]` and `[1, 2]` are identical types, but in C++, they
are different types)

Generated code for `TensorFactoryTest.TensorConstruct`:

```CUDA
__global__ void kernel1(int64_t i0, int64_t i1, int64_t i2, int64_t i3, Tensor<int64_t, 2, 2> T0) {
  NVFUSER_DEFINE_MAGIC_ZERO;
  Array<int64_t, 2, 1> a4;
  a4 = {i2, i3};
  Array<int64_t, 2, 1> a5;
  a5 = {i0, i1};
  Array<Array<int64_t, 2, 1>, 2, 1> a6;
  a6 = {a5, a4};
  #pragma unroll
  for(nvfuser_index_t i7 = 0; i7 < 2; ++i7) {
    int64_t i8;
    i8 = 2 * i7;
    Array<int64_t, 2, 1> a9;
    a9 = a6[i7];
    #pragma unroll
    for(nvfuser_index_t i10 = 0; i10 < 2; ++i10) {
      int64_t i11;
      i11 = i10 + nvfuser_zero;
      T0[(i8 + i11)] = (a9[i11]);
    }
  }
  NVFUSER_UPDATE_MAGIC_ZERO;
}
```