Allocate intermediate tensors using the shape provided by Allocate node (#527)
Some of the shift tests use global memory tensors just for testing the
functionality. In those cases, there are intermediate global tensors
that are extended with halo. Their allocation sizes are not the same as
the logical size represented by their allocation domains, but slightly
larger to account for the halo regions.

With the current main, the intermediate tensors for `FusionShift2` are:

```
Intermediate global buffers:
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
```

This PR effectively reverts the change of executor.cpp by #324. I think
this is the easiest fix.

With this fix, they are:

```
Intermediate global buffers:
  Float [10, 11] is_zero_initialized: 0
  Float [11, 12] is_zero_initialized: 0
  Float [11, 12] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
  Float [9, 11] is_zero_initialized: 0
```

Alternatively, we may also want to think about extending the allocation
domain itself such that it truly holds the allocation information, not
just the logical domains and their ordering but their actual sizes.
Given that the halo support is experimental and likely needs an almost
complete rewrite, a quick fix would be just fine for now.