Make comparison op return bool (#522)
This PR is stacked on https://github.com/NVIDIA/Fuser/pull/520

In C++, although we are allowed to overload comparison op to return an
arbitrary type, it commonly just returns a `bool`. It is commonly
assumed that comparison op will return `bool`, especially, in C++
standard library, such assumption also exist. For example,
`std::vector<T>{} < std::vector<T>{}`, returns a `bool` instead of
`decltype(T{} < T{})`. Due to this assumption, it actually makes more
sense that we also return `bool`, instead of `DynamicType` for these
ops.

Changing this means we will be unable to support `T`s whose `T < T` does
not return a `bool` well, but in my entire career, I have never seen any
of such case. The benefit of this switch is, we can now more seamlessly
integrate with the oter parts of C++.