Concretize in topological order during mutation.
Previously we concretized all reshapes then all slices then all resizes.
This failed the FusionSliceForNanoGPT3_CUDA test which has the
slice->reshape->slice->reshape pattern, since we were concretizing
reshapes that had inputs which were not yet concretized and which
actually needed to be replaced. Since concretization of slices and
reshapes can actually replace a TensorView, we need to ensure that those
are done in the correct order.