Use deque instead of stack.
We remove from the front and push to the back to process Statements in
FIFO order. This ensures we have traversed in a reverse topological
order, so that we can safely remove any TensorViews downstream of the
Statement we're looking at, as they have already been processed and
should not appear later in the stack (though we should check still
because they might be an output).