Skip swizzle in target for transform replay (#547)
When doing transform replay, we should skip swizzle in the target, not
erase it. cc: @liqiangxl

Background: To support shared memory epilogue for matmul, we need to
schedule the mma output and propagate to the smem tensor. The smem
tensor should already be swizzled before the propagation, and the
propagation should respect that swizzle. This PR will be tested in
https://github.com/NVIDIA/Fuser/pull/387