Basic naive bmm schedule (#577)
This PR modifies `scheduleMatmul` to support basic bmm by parallelizing
the batch dim on BIDz. I don't know how good is the perf for this
schedule, but this is not the purpose of this PR. The main purpose for
this PR is to have something functionally working, so that I can unblock
my next step for adding a bmm scheduler to the registry which also
dispatch to `scheduleMatmul` for the separate kernel split-k. I will
leave it to @drzejan2 to implement a performant bmm.