Allow `ExpressionEvaluator` to eagerly evaluate tensor expressions (#572)
Refactor current tensor metadata support to allow `ExpressionEvaluator`
to eagerly evaluate tensor expressions. With all the infrastructure we
have built, adding support for eagerly evaluate tensor is pretty easy
and straightforward, so why not support it?

Currently, only C++ operators on `at::Tensor` is supported, if we
continue to invest on this direction to add support for more tensor ops,
this would at least make it much easier to check our codegen with aten
when writing tests because we no longer need to manually write out the
aten implementation.

In the future, with the planned `inputs` vs `kernel_inputs` for
`kir::Kernel`, this enables the possibility for us to easily offload
some computation from our kernel to ATen, especially those ops that we
feel difficulty to handle but is free for eager mode (for example,
reshape on fusion inputs).

Also, if we built out `ExpressionEvaluator` for all tensor ops, we can
deprecate
https://github.com/NVIDIA/Fuser/blob/main/tools/cpp-repro-gen.py, and
instead, add the capability for our python frontend to dump fusion
definitions as C++ code. In the past, we discussed about adding this
capability with Kevin, Jie, and Christian, but we preferred
`cpp-repro-gen.py` because it can generate ATen repro.