Allow linking UCC in nvfuser build (#581)
After this PR, nvfuser build may link with UCC.

For submodule build, no change is needed. If the pytorch build has UCC
enabled, nvfuser build will link with UCC, too.

For standalone top level build, UCC is **not** enabled in nvfuser by
default. User will need to specify `--build-with-ucc` with setup.py to
enable build with UCC. This only happens if pytorch is built with UCC as
well.

This PR also fixed a minor issue in C++ definitions of `ProcessGroupUCC`
from pytorch.

Related #483