Set extents of empty tensors to zeroVal() during concretization (#449)
A number of issues have come up when trying to process empty tensors
(i.e. ones with at least one non-reduction axis with extent of zero)
during scheduling and lowering. See: #264 #369 #269. Additionally, we
now assume extents are positive (#440). Along with #543, this PR makes
that a reality by removing all intermediate empty tensors.

This PR:
- Marks a `Fusion` as dynamic if dynamic reshapes/resizes exist or if
_any_ alive `TensorView` has a static size-zero extent or a dynamic
extent, since it might be empty. **This is means only Fusions with
nothing but concrete non-zero sizes are static now.** That is, even if
all static shapes are provided, it will be marked as a dynamic Fusion
and those `TensorView`s will be modified during concretization.
- Adds a pass done during `getConcretizationInfo()` that collects a
vector of empty tensors which are not fusion inputs. It does not
traverse their definitions, since there is nothing to compute for an
empty tensor.
- During concretization, sets the size-0 extents of identified empty
tensors to constant 0.

When encountering a new set of input sizes/scalars, we evaluate a
minimal set of `Val`s (those that appear in dynamic extents), and only
proceed with removing branches if any of these are zero. So there is a
rather quick path to re-using concretizations in the common case where
none of the extents are zero.

Even after #543, this PR does not guarantee that all tensors present in
the Fusion during scheduling have non-zero extent. It does guarantee
that any remaining empty tensors are either fusion inputs or outputs,
and that empty tensors will have constant 0 extents in any empty
dimensions. Stripping empty inputs and outputs from the Fusion could
potentially be done at segmentation, but should only be done if it does
not result in additional kernels being launched; that is left for
another PR (see #448).

Fixes #365 and fixes #264. This replaces PRs #369 and #269.

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>