Add NVFUSER_DISABLE=kernel_reuse option (#563)
This adds an option to avoid re-using compiled `FusionKernelRuntime`s
inside `FusionExecutorCache` when encountering new input shapes. This
will force segmentation/compilation to repeat for every new set of input
shapes. This is desireable for benchmarking, since it ensures repeatable
performance independent of the order of inputs. Note that kernels will
always be re-used when the input shapes are repeated, and this PR does
not change the fast path to reuse in that case.