Use expanded extent for empty check in RemoveEmptyPass (#604)
When replacing empty tensors with `full` in `RemoveEmptyPass` (#543), we
use the `extent` to determine the output shape. This is a problem if
that tensor was the output of an `expand` call, because then even a
tensor that is not empty might appear to be so if an expanded extent is
zero. This change addresses this issue by using `getMaybeExpandedExtent`
in `RemoveEmptyPass` and during concretization for finding empty
tensors.

Fixes #603