Allow struct dtype to be named (#647)
If a struct is already defined in the `runtime/`, we do not have to
generate its definition during lowering. For that case, we just generate
the name as defined in `runtime`. For example, we do not have to
generate tensor metadata as `struct {....}`, instead, we can just do
things like `Tensor<float, 4, 4>`.