Initial `Kernel::parameters()` support (#644)
This PR adds `Kernel::parameters()`, which currently can only be
`[*inputs, *outputs, *intermediates]`, in the future, it will contain
all parameters, including special items such as Philox seeds for RNG,
and TMA descriptors.

This PR also adds `GpuLower::allKnownVals()`, which currently can only
be `[*inputs]`. In the future, `allKnownVals()` will contain values
hoisted to ATen or host computation, for example, it can hoist a
metadata-only operation such as transpose/view into the ATen, or hoist
part of the index to host.

The executor logic is also changed to adopt `Kernel::parameters()`. With
the change, it is like:
```C++
bind inputs
allocate outputs
bind outputs
allocate intermediates
bind intermediates
use expr evaluator to evaluate the entire Kernel::parameters(), and put the result into a buffer
append philox to the buffer if needed
launch kernel with the buffer
```

With this change, the previous framework based on `KernelArgumentHolder`
and `ArgAbstract` is now deprecated. It will be removed in followup PRs.

This PR adds a total of 200+ lines of code to nvFuser. But with the
planned cleanups, I would expect to see a net reduction in code when
everything is done.