Cache metadata access expr inside IR container (#648)
`GetMetaData` expr is so widely used. We should avoid creating many
instances of it. Also, this would be helpful for kernel parameters
support. For example, if we set `Kernel::parameters` as `v1 =
GetMetaData(tv1)` instead of `tv1`, then without this PR, the generated
kernel might not work because in the tensor index, we might have `v2 =
GetMetaData(tv1)`, although `v2` and `v1` are the same, if scalar
hoisting does not work for some reason and `v2` is not replaced with
`v1`, the lowered expr might fail, because `tv1` is not a kernel
parameter, therefore unknown to the kernel, therefore `v2` can not be
evaluated in the kernel.