Allow `Expr::evaluate` to have full access to the evaluator (#650)
Some expressions might need context information to evaluate, for example
`GetMetaData` might need the split size for symbolic splits in order to
compute the sizes and strides of the allocation domain. So I added a
const reference of the expr evaluator to `Expr::evaluate`. I think it
makes sense to have access to such context information. This change will
also allow expr evaluator to evaluate tensor factories like `FullOp` on
the non-zero device, because the expr evaluator can just store the
desired device and the `FullOp::evaluate` can pull that information out.
However, `ExpressionEvaluator` must be passed as a const reference to
ensure that the `evaluate` function does not modify it. For example, I
don't want someone to do a `ee.bind(...)` inside the `evaluate`
function, because this definitely looks too hacky to accept.