Refactor RNG (#655)
This PR remove the special handling of RNG in codegen and executor. With
this PR, if you want to do a stochastic RNG, you can just create an
`RNGOp` without specifying its seed and offset. During lowering,
`assignRNGOffset` will complete the missing seed and offset by
generating one using `GetRNGSeedAndOffsetFromHost` and append the
outputs of `GetRNGSeedAndOffsetFromHost` to `kernel()->parameters()`, so
that the fusion executor will automatically evaluate
`GetRNGSeedAndOffsetFromHost` which will obtain a seed and offset from
ATen.