Use `GetMetaData` for stride computation (#649)
The definition of `Tensor` in runtime has been changed as

```CUDA
template <typename T, int Dims, int AllocDims = Dims>
struct Tensor {
  __device__ T& operator[](nvfuser_index_t ind) {
    return data[ind];
  };

  T* data;
  Array<nvfuser_index_t, Dims, 1> logical_size;
  Array<nvfuser_index_t, AllocDims, 1> alloc_stride;
};
```

That is, we are more explicit about whether we are referring to the
size/stride of allocation domain or rFactor domain. On the host, the
`PolymorphicValue` of tensor metadata now stores all the four
`logical_size`, `logical_stride`, `alloc_size`, `alloc_stride`.

The utility that converts sizes and strides wrt rFactor to allocation
domain was `inferAndValidateAllocationSizesAndStrides`, this function
has been moved to `tensor_metadata.cpp` and is now private to that file.
The new way to compute sizes and strides of allocation domain is:

```C++
expression_evaluator.evaluate(IrBuilder::metadataExpr(tv));
```

I need to change `SchedulerRuntimeInfo::SchedulerRuntimeInfo`,
`getKernelArgument`, and `validateAlignedVectorizedFusionInputOutput` to
use this new approach.

With this change, `getTensorArg` and `KernelArgumentHolder::getBuffer`
are no longer used. We should be ready to remove all the subclasses of
`ArgAbstract`, but I am not doing this clean up. These cleanups will be
left as next PR.