Fix vectorization with reduction domains in fusion inputs (#659)
Issue #658 is because we are using the number of consumer domains to
compute a position in its producer tensor. Since a producer may have
reduction IDs that the consumer do not have, the computed position can
result in larger than intended, making fewer number of vectorized
innermost domains. This issue happens only with the reduction and
normalization schedulers.

Fixes #658