Alias allocations with same size elements despite different dtypes (#665)
Currently the `reuseMemoryAllocations` lowering pass reuses
`TensorView`s if they have compatible shapes and vectorization, and the
same dtype. This PR replaces the dtype condition with a check that the
size of each element is the same.

It is somewhat rare in our test suite that we find opportunities to
re-use one buffer with another of a different type, but it does happen.
For example, one case in `FusionWelfordShmoo_CUDA` corresponding to
`testWelford(DataType::ComplexFloat, 1, 320, 256);` currently shows:
```
ptxas info    : Used 61 registers, 16 bytes smem, 464 bytes cmem[0]
```
With the current PR, this becomes
```
ptxas info    : Used 59 registers, 16 bytes smem, 464 bytes cmem[0]
```
since we re-use two register `TensorView`s that change from `int64_t` to
`std::complex<float>`.

Re-using shared memory in these cases is likely always beneficial, but
the effect of explicitly re-using registers is hard to analyze since the
compiler pipeline is so complicated. It is likely rarely harmful, so
this behavior is on by default. However, it can be disabled using
`DisableOption::ReuseMismatchedTypeRegisters` or
`NVFUSER_DISABLE=reuse_mismatched_type_registers`.