Split Not, And, and Or into logical and bitwise versions (#691)
Currently we have `UnaryOpType::Not` which implements both `!` and `~`
depending on input types, and similar for `BinaryOpType::And` and
`BinaryOpType::Or`, which implement `&&`/`&` and `||`/`|`. This can lead
to unintuitive bugs when types are implicitly cast between `int` and
`bool` (see #688). It is much clearer to split these operators, which
this PR does. In summary, we remove the following operation types:
- `UnaryOpType::Not`
- `BinaryOpType::And`
- `BinaryOpType::Or`
- `BinaryOpType::Xor`

and we introduce these new op types:
- `UnaryOpType::LogicalNot`
- `UnaryOpType::BitwiseNot`
- `BinaryOpType::LogicalAnd`
- `BinaryOpType::BitwiseAnd`
- `BinaryOpType::LogicalOr`
- `BinaryOpType::BitwiseOr`
- `BinaryOpType::BitwiseXor`

We also introduce the following python ops:
- `fd.ops.logical_not`
- `fd.ops.logical_and`
- `fd.ops.logical_or`

Note that we do not implement `LogicalXor` which for boolean inputs is
simply `x != y`.

Each of the bitwise operations, if receiving all `bool` inputs, will
convert to the equivalent logical operation which will return a `bool`.
If any argument is int then both arguments are cast to int and an int is
returned.

Each of the logical operations, if receiving non-`bool` inputs, will
first cast arguments to `bool`.

This PR also bumps the version since we are making a minor change to the
python interface.


This PR replaces #688.