Give dynamic shared buffers address expressions (#689)
This is related to #221. Currently, dynamic shared memory is used in two
ways: either implicitly by reductions or when we have a `TensorView`
with `MemoryType::Shared`. Note that our IR does not explicitly
represent reduction (or Welford) smem at all. I'll refer to the shared
memory space reserved for reductions and welfords as "reduction smem"
and to all the shared mem above it via `TensorView`s as `tv smem`.

Currently when we generate CUDA kernels, we look for reductions and
Welford ops and find the largest data type and we reserve that amount
times the block size at the beginning of the dynamic smem array
`shared_mem`. We then set `smem_offset` to that size. When `tv smem` is
defined in the kernel, we align the `smem_offset` address to 16 bytes
and use it for the buffer definition, then we add the new buffer's size
to `smem_offset` in the generated kernel.

This method makes memory re-use cumbersome. We currently support
aliasing identically-sized buffers, but if we wanted to create a new
allocation with a different size between previous memory locations, we
could not do that since we don't represent the smem offsets anywhere.

This PR creates an `address()` attribute on `kir::Allocate` which holds
a scalar expression for the number of bytes above `smem_offset` at which
to start the allocation. When generating a kernel, we use these
expressions to directly find the address of the new buffer, and we never
modify `smem_offset`. During lowering, in the `reuseMemoryAllocations`
pass, we assign those address `Val` attributes without re-using any `tv
smem`.

Note that this PR could potentially increase register use since we don't
do any explicit CSE on these expressions: if there are multiple of smem
TVs their offset expressions can grow in size due to intermediate
alignment expressions. However, I didn't observe any change in register
usage in our current tests.