Fix scheduling with reshape and resize (#710)
Stacked on top of #706. Part of work to enable slice vectorization.

This is a small bug fix for fusions with both reshape and resize.
`propagateViewTransforms` (renamed to `propagateReshapeTransforms`) was
attempting to propagate reshape as it's also an rfactor op, which is
obviously not necessary. The added tests fail with this fix.