Randomize `Val*` for expr simplifier test (#711)
Currently, in expr simplifier tests, we always create a new `Val*` for
each variable that appears in the expression. For example, in `"i + i +
i"_`, we will create three `NamedScalar`s, all with name `i`. This
behavior does not have good coverage on the case where there are the
same `Val*` as different inputs of the same expr. And there can be bugs
related to this, for example, if I want to simplify `FlattenedAdd(i, i,
i, -i)`, if not implemented carefully, I might be removing all the `i`s
in the `FlattenedAdd`, instead of just removing one, if all the `i`s has
the same `Val*`. In this PR, I changed the "stupid_simple_compiler" to
randomly decide whether to reuse an existing `Val*` or to create a new
one. This will increase our test coverage, although determinism might
get hurt.