Use `DataType::Index` whereever it should (#696)
This PR is stacked on https://github.com/NVIDIA/Fuser/pull/700. As the
title describe, this PR makes our system use `DataType::Index` instead
of `DataType::Int` whereever it should. In principle, we should have
been using `DataType::Index`, but due to historical reasons, we are not
doing so.

This start to cause trouble as I am working on using structured tensor
size (https://github.com/NVIDIA/Fuser/pull/684), because the sizes in
tensor metadata is `DataType::Index` but we are using `DataType::Int`
for our ID extents, and our IR mutator is not able to replace a `Val`
with another `Val` of different dtype. As a result, the
`replaceSymbolicSizes` will fail during lowering.

To resolve the above issue, we should fix our integer usage, and use
`DataType::Index` whereever it makes sense. In this PR I added asserts
on the ctor of `IterDomain` and `kir::ForLoop` to check that the `Val`s
used to construct them are `DataType::Index`. And our system is updated
to generate indices with `DataType::Index` so that these newly added
checks pass.

Because of the `resolveIndexDtype` call in `scalar_hoist.cpp` (which was
used to workaround the integer dtype issue by hacking the dtype of
scalar computation in kernel), this PR should have very change in the
final SASS. So I didn't check benchmark results (it takes really long).
However, in my next PR, I will remove `resolveIndexDtype` and compare
the perf of benchmarks in that PR.