Do not allow non-projected extents (#714)
It seems odd to me that
`ContiguousInnerDimensionMapper::getProjectedExtent` automatically
returns 1 for any domain that is not projected. It seems that was
necessary for it to work under the non-recording mode, but there's no
reason to use the function if no recording projected extents.

This PR makes `getProjectedExtent` with non-projected domains result in
an error. Also changed `distributePE` and `combinePE` to do nothing
while not recording as they use `getProjectedExtent`.

Also added an error check to `addProjectedExtent`.