More robustly detect size one dim (#719)
Fixes https://github.com/NVIDIA/Fuser/issues/718. Currently, we use
`isOneInt` to detect if a dimension is size 1. This is not robust,
because not all size one dims are constant one. A dim's extent might be
something like `(nvfuser_index_t)1L`, which is still one. For that case,
we should be able to detect it correctly.